feat(commitlint): enforce 50/72 formatting
Enforce a maximum line length of 72 characters for the entire message,
and of 50 for the subject. Fail linting if the message does not satisfy
the requirements.

The reasoning behind allowing the header to have a length of 72
characters is to compensate for the reduced space that the subject line
has when taking into account commit type and scope.