fix(xorg): keep touchpad tapping enabled indefinitely
Add an Xorg configuration file `90-touchpad.conf` configuring input
behaviour for touchpads using an `"InputClass"` section, and set
`Option "Tapping"` to `"on"` in order to keep tapping enabled
indefinitely.

This is necessary because the default behaviour in Xorg seems to be to
stop recognizing tapping on touchpads after a certain period of time.