feat(tmux): use current directory for default session
Use the current directory as the working directory for the default
session that the client attaches to when running `tmux` without any
arguments. This replaces `home` as the default session. The session
`home`, which uses the current user's home directory as its working
directory, is still created, but not attached to the client.