feat(wezterm)!: remove `tmux-sessionizer` integration
Remove keybinding `CTRL|SHIFT|p` that runs `tmux-sessionizer` when
inside a shell. This can also be achieved by manually typing in the
command in the shell, making the binding redundant, and since the
binding only works in a shell, it is also brittle. This means it is
cleaner and more predictable to remove the binding.

BREAKING CHANGE: This change removes the keybinding `CTRL|SHIFT|p`.