chore!: move local config files into separate directory
- Move local config files in `src` into a newly created `src/local`
  directory and update paths in `README.md`
- Update ignore paths in `cspell.json` to refer to `src/local` where it
  is necessary
- Move `.gitignore` into `src/local/.config` and update the ignore rules
  to be relative to this directory

This separates the local config files from the global configuration
files in a clear way, which increases maintainability. Moving
`.gitignore` means that the ignore rules can be simplified.

BREAKING CHANGE: This commit modifies the repository structure, and
`stow` will be unable to reconcile the differences automatically. Manual
intervention is required.